Refactor userReducer
