+++ src/redux/store.js
@@ -2,11 +2,11 @@ import { applyMiddleware, combineReducers, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
 import bookDoctorReducer from './bookDoctor/doctorActions';
-import signupReducer from './signup/signup';
+import userReducer from './user/userActions';
 
 const reducer = combineReducers({
   bookDoctorReducer,
-  signupReducer,
+  userReducer,
 });
 
 // creating a store
